Workaround: don't include source for udebs in images
* list2cds: set "Component" for udebs to "debian-installer
* merge_package_lists: skip udebs when including source

When building images for Etch+1/2, which takes regular packages from
stable and udebs from testing, there is a problem when adding source
for udebs as existing code will take source for both debs and udebs
from stable. This can result in either incorrect source being included
(version mismatch) or errors (new udebs for which no match can be found
in the Source.gz file that is used).

A proper solution would require that the debian-installer section gets
its own Sources.gz file, but that is currently just not available.
After trying some things the conclusion was that the best solution for
now is to just not include source for udebs at all.

Patch written by Steve McIntyre.