Patch from Ian Campbell <ijc@hellion.org.uk>:
Detect duplicates in the extra images used for installation

Several files included under install.{386,amd} are actually identical to
others. Rather than duplicating them attempt to detect this and hardlink
them.

Since images can be downloaded from the web at build time it is not
always possible to detect identical files by looking for the symlinks
created by the debian-installer build. Instead we pass a list of
potential "doppelgangers" to extra_image each of which is checked for
similarity to the new image.

I added gtk/vmlinuz (which is always the same as plain vmlinuz) which
although not necessary makes it more explicit which kernel goes with the
initrd at very little cost.

Hardlinks are used in preference to symlinks since these are expected to
work better with isolinux and the Xen tools.