Bugfix: fix the remove_Packages_entry() helper function
This was a bug introduced with the backports support. Lots of the
internal helper functions take 4 arguments now: ($dir, $arch,
$in_backports, $package_info). That's fine, but I forgot to update the
code in remove_Packages_entry() to deal with the $in_backports
argument.

This led to the md5sums.txt files in multi-disc sets being broken -
the last file(s) removed on a given disc would still be listed in the
md5sums.txt file, even though the file itself had been removed.