mydistro: Make sure package is removed from the list on failure
If add_missing() fails we remove package from the list unless this is
soft dependency and then add it back despite of failure.

On the other hand add_missing() restores original list on failure where
package already pushed at the end. Thus there is no reason to call
remove_entry() followed by push to re-add it at the end of the list.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>