mydistro: Take a copy of hash on dereference only when necessary
It is performance costly operation and thus should be avoided as much
as possible. Do it only when original hash being modified or passed to
a function that modifies or enrolls it in other data structure.

Defer copy when it is not possible to avoid it completely.

In add_missing() keep $pkgname untouched as $pkgin is unmodified.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>