mydistro: Check requested version of package available as well
Before this we check available package version only for dependencies
not package itself. This isn't correct as first package listed in OR
dependency might not be available or not acceptable version skipping
further candidates in list.

Behaviour was observed during Debian GNU/Linux 7 (wheezy) build with
libio-socket-ip-perl whose Depends tag is:

   perl, netbase, perl (>= 5.15.6) | libsocket-perl (>= 1.97)

but perl version is 5.14.x and thus libsocket-perl should be choosen
as it's version 2.002.x satisfies condition >= 1.97.

While there check that package exists before accessing it's keys.

Fixes: 9fa8de1b10ac ("mydistro: Fix perl deep recursion in add_missing() and speedup things")
Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>